@@ -6,53 +6,53 @@
   "nativescript": {
     "id": "org.nativescript.taxi",
     "tns-android": {
-      "version": "3.3.1"
+      "version": "3.4.1"
     }
   },
   "dependencies": {
-    "@angular/animations": "~4.4.5",
-    "@angular/common": "~4.4.5",
-    "@angular/compiler": "~4.4.5",
-    "@angular/core": "~4.4.5",
-    "@angular/forms": "~4.4.5",
-    "@angular/http": "~4.4.5",
-    "@angular/platform-browser": "~4.4.5",
-    "@angular/platform-browser-dynamic": "~4.4.5",
-    "@angular/router": "~4.4.5",
-    "@ngrx/effects": "^4.0.5",
-    "@ngrx/store": "^4.0.3",
+    "@angular/animations": "~5.0.0",
+    "@angular/common": "~5.0.0",
+    "@angular/compiler": "~5.0.0",
+    "@angular/core": "~5.0.0",
+    "@angular/forms": "~5.0.0",
+    "@angular/http": "~5.0.0",
+    "@angular/platform-browser": "~5.0.0",
+    "@angular/platform-browser-dynamic": "~5.0.0",
+    "@angular/router": "~5.0.0",
+    "@ngrx/effects": "^4.1.1",
+    "@ngrx/store": "^4.1.1",
     "nativescript-android-utils": "^1.0.2",
-    "nativescript-angular": "~4.4.1",
+    "nativescript-angular": "~5.0.0",
     "nativescript-appversion": "^1.4.1",
-    "nativescript-cardview": "^2.0.2",
     "nativescript-geolocation": "^4.2.0",
     "nativescript-localstorage": "^1.1.5",
     "nativescript-theme-core": "~1.0.4",
     "reflect-metadata": "~0.1.8",
-    "rxjs": "~5.4.3",
-    "tns-core-modules": "^3.3.0",
-    "zone.js": "^0.8.16"
+    "rxjs": "^5.5.0",
+    "tns-core-modules": "^3.4.0",
+    "zone.js": "^0.8.4"
   },
   "devDependencies": {
-    "@angular/compiler-cli": "~4.4.5",
-    "@ngtools/webpack": "~1.6.0",
+    "@angular/compiler-cli": "~5.0.0",
+    "@ngtools/webpack": "~1.8.2",
     "@types/lodash": "^4.14.70",
     "babel-traverse": "6.25.0",
     "babel-types": "6.25.0",
     "babylon": "6.17.4",
     "copy-webpack-plugin": "~4.0.1",
+    "css-loader": "~0.28.7",
     "extract-text-webpack-plugin": "~3.0.0",
     "lazy": "1.0.11",
     "nativescript-css-loader": "~0.26.0",
-    "nativescript-dev-typescript": "~0.5.0",
-    "nativescript-dev-webpack": "^0.8.0",
+    "nativescript-dev-typescript": "0.6.0",
+    "nativescript-dev-webpack": "0.9.0",
     "nativescript-worker-loader": "~0.8.1",
     "raw-loader": "~0.5.1",
     "resolve-url-loader": "~2.1.0",
     "tslint": "^4.5.1",
-    "typescript": "~2.3.4",
+    "typescript": "~2.4.2",
     "uglifyjs-webpack-plugin": "^1.0.0",
-    "webpack": "~3.2.0",
+    "webpack": "~3.8.1",
     "webpack-bundle-analyzer": "^2.8.2",
     "webpack-sources": "~1.0.1"
   },
@@ -60,10 +60,10 @@
     "ns-bundle": "ns-bundle",
     "publish-ios-bundle": "npm run ns-bundle --ios --publish-app",
     "generate-android-snapshot": "generate-android-snapshot --targetArchs arm,arm64,ia32 --install",
+    "tslint": "tslint --project tsconfig.json --config tslint.json",
     "start-android-bundle": "npm run ns-bundle --android --run-app",
     "start-ios-bundle": "npm run ns-bundle --ios --run-app",
     "build-android-bundle": "npm run ns-bundle --android --build-app",
-    "build-ios-bundle": "npm run ns-bundle --ios --build-app",
-    "tslint": "tslint --project tsconfig.json --config tslint.json"
+    "build-ios-bundle": "npm run ns-bundle --ios --build-app"
   }
 }
