@@ -1,64 +1,20 @@
-import { Injectable } from "@angular/core";
-import { Http, ConnectionBackend, RequestOptions, Request, RequestOptionsArgs, Response, Headers } from "@angular/http";
-import { Observable } from "rxjs/Observable";
-import "rxjs/add/operator/map";
-import "rxjs/add/operator/catch";
+import { Injectable } from '@angular/core';
+import { HttpEvent, HttpInterceptor, HttpHandler, HttpRequest } from '@angular/common/http';
 import { Config } from "../config";
+import { Observable } from "rxjs/Observable";
 
 @Injectable()
-export class InterceptedHttp extends Http {
-    constructor (backend: ConnectionBackend, defaultOptions: RequestOptions) {
-        super(backend, defaultOptions);
+export class InterceptedHttp implements HttpInterceptor {
+    intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
+
+        let token = localStorage.getItem("token");
+
+        const authReq = req.clone({
+            setHeaders: {'Content-Type': 'application/json', 'x-driver-auth': token ? token : ''},
+            url:  `${Config.APIURL}${req.url}`
+        });
+
+        return next.handle(authReq);
     }
 
-    request(url: string | Request, options?: RequestOptionsArgs): Observable<Response> {
-        return super.request(url, options).catch(this.catchAuthError(this));
-    }
-
-    get(url: string, options?: RequestOptionsArgs): Observable<Response> {
-        url = this.updateUrl(url);
-        return super.get(url, this.getRequestOptionArgs(options));
-    }
-
-    post(url: string, body: string, options?: RequestOptionsArgs): Observable<Response> {
-        url = this.updateUrl(url);
-        return super.post(url, body, this.getRequestOptionArgs(options));
-    }
-
-    put(url: string, body: string, options?: RequestOptionsArgs): Observable<Response> {
-        url = this.updateUrl(url);
-        return super.put(url, body, this.getRequestOptionArgs(options));
-    }
-
-    delete(url: string, options?: RequestOptionsArgs): Observable<Response> {
-        url = this.updateUrl(url);
-        return super.delete(url, this.getRequestOptionArgs(options));
-    }
-
-    private updateUrl(req: string) {
-
-        return Config.APIURL + req ;
-    }
-
-    private getRequestOptionArgs(options?: RequestOptionsArgs): RequestOptionsArgs {
-        if (options == null) {
-            options = new RequestOptions();
-        }
-        if (options.headers == null) {
-            options.headers = new Headers();
-        }
-        options.headers.append("Content-Type", "application/json");
-        options.headers.append("X-Driver-Auth", localStorage.getItem("token"));
-
-        return options;
-    }
-
-    private catchAuthError (self: InterceptedHttp) {
-        return (res: Response) => {
-            if (res.status === 401 || res.status === 403) {
-                console.log(res);
-            }
-            return Observable.throw(res);
-        };
-    }
 }
