@@ -6,9 +6,9 @@ import { Router } from "@angular/router";
 
 import { IUser } from "../../shared/models/user/user.model";
 import { NUser } from "../../modules/state-managment/actions/user.action";
-import { Store } from "@ngrx/store";
+// import { Store } from "@ngrx/store";
 import { IAppState } from "../../modules/ngrx/index";
-import { Observable } from "rxjs";
+import { Observable } from "rxjs/Observable";
 import * as dialogs from "ui/dialogs";
 declare var android: any;
 import * as application from "tns-core-modules/application";
@@ -28,23 +28,23 @@ export class CardInfoComponent implements OnInit {
 
     constructor(private page: Page,
                 private router: Router,
-                private store: Store<IAppState>) {
+               /* private store: Store<IAppState>*/) {
 
          this.user = new User();
-         this.user$ = this.store.select("user");
+         // this.user$ = this.store.select("user");
 
-         this.user$.subscribe(e => {
-             if (e) {
-                 this.isLoading = false;
-                 this.user.balance = Number(e.balance).toFixed(2);
-                 this.user.signal = e.signal;
-                 this.user.name = e.name;
-             }
-         }, e =>  dialogs.alert({
-             title: Config.messages.error.title,
-             message: e.message,
-             okButtonText: Config.messages.button.ok
-         }));
+         // this.user$.subscribe(e => {
+         //     if (e) {
+         //         this.isLoading = false;
+         //         this.user.balance = Number(e.balance).toFixed(2);
+         //         this.user.signal = e.signal;
+         //         this.user.name = e.name;
+         //     }
+         // }, e =>  dialogs.alert({
+         //     title: Config.messages.error.title,
+         //     message: e.message,
+         //     okButtonText: Config.messages.button.ok
+         // }));
     }
 
     goToSettings() {
@@ -64,25 +64,25 @@ export class CardInfoComponent implements OnInit {
 
     ngOnInit() {
 
-        this.store.dispatch(new NUser.LoadAction()); /* this is dispatcher for loading user */
+        /*this.store.dispatch(new NUser.LoadAction());*/ /* this is dispatcher for loading user */
 
         this.page.androidStatusBarBackground = Config.actionBarColor;
 
-        isEnabled()
-            .then(hasAccess => {
-                if (!hasAccess) {
-                    enableLocationRequest()
-                        .then(() => {
-                            this.startBackgroundService();
-                        }, (e) => {
-                            this.startBackgroundService();
-                            console.log("Error Decline Request: " + (e.message || e));
-                        });
-                } else {
-                    this.startBackgroundService();
-                }
-            }, (e) => {
-                console.log("Error is Enabled: " + (e.message || e));
-            });
+        // isEnabled()
+        //     .then(hasAccess => {
+        //         if (!hasAccess) {
+        //             enableLocationRequest()
+        //                 .then(() => {
+        //                     this.startBackgroundService();
+        //                 }, (e) => {
+        //                     this.startBackgroundService();
+        //                     console.log("Error Decline Request: " + (e.message || e));
+        //                 });
+        //         } else {
+        //             this.startBackgroundService();
+        //         }
+        //     }, (e) => {
+        //         console.log("Error is Enabled: " + (e.message || e));
+        //     });
     }
 }
